Implement selection switching and scrolling to menu
Add ability for the selection to be changed programmatically. To enable
this, added a new feature into each menu implementation which handles
any keybinds which were not handled globally. This required a rewrite of
some menu handlers. This will be called the "menu-local keybinds"
system.

Added this ability to the menu component (after figuring out the math)
and programmed it in to the list component.