Guess downloads dir exclusively from queue
This removes the requirement to "parse" the newsboat config file (if you
could call whatever we were doing "parsing").

Now, the longest common substring to all the paths from all entries will
be used for the downloads dir. This will soon be expanded to take a
value from the new config file, which overrides this guessed value.