Fix very insidious queue reload bug
This took me almost two hours to locate.

When we reload the queue, we completely reallocate. However, Go will
detect concrete references to the old buffer through the linkmap. So,
all the pointers to the original buffer will be retained along with the
buffer, leading code using one set of functions to modify the now
unreachable copy of each episode and another set of code reading from
the newly created copy via the newly reallocated buffer.

This is really bad because it caused data loss in a load of non obvious
circumstances.

Moral of the story: don't rely on pointers from buffers which you intend
to re-allocate - even in memory safe languages.

Fixed by instead taking the address of the episode we created on the
local stack of the parsing routine and using that copy as an
authoritative copy. This will escape to the heap and avoid being
reallocated for the lifetime of the program.